+++ onnxruntime/src/gen/java/org/bytedeco/onnxruntime/RunOptions.java
@@ -6,6 +6,9 @@ import java.nio.*;
 import org.bytedeco.javacpp.*;
 import org.bytedeco.javacpp.annotation.*;
 
+import org.bytedeco.dnnl.*;
+import static org.bytedeco.dnnl.global.dnnl.*;
+
 import static org.bytedeco.onnxruntime.global.onnxruntime.*;
 
 
@@ -28,7 +31,7 @@ public class RunOptions extends BaseRunOptions {
   public native int GetRunLogVerbosityLevel();
 
   public native @ByRef RunOptions SetRunLogSeverityLevel(int arg0);
-  public native int GetRunLogSeverityLevel();
+  
 
   public native @ByRef RunOptions SetRunTag(@Cast("const char*") BytePointer run_tag);
   public native @ByRef RunOptions SetRunTag(String run_tag);